requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
numpy>=1.24.0
openai>=1.35.0
